Service Layer Skeleton
Began work on the service layer.
The methods are in place with the correct annotations.
The properties file has been updated with the XML string formats that
can be used in the service layer to return well-formed XML, however Team
Win is still going to seek alternatives to this because XML marshaling
and unmarshaling SHOULD be an easy thing to implement seeing that it is
not a new concept.

If this fails, we will resort to the properties file for forming xml
responses and manually parsing incoming XML to create objects...