XStream
Started writing the MessageBodyWriters that will ultimately marshal and
unmarshal objects to and from XML.
Also, it has come to our attention that according to the spec, a new or
updated appointment may have lab tests included in it. The business
layer will need to be updated accordingly.

I also made an attempt to use ivy for dependency management but
unfortunately I have no idea where some of those libraries are so I'm
saving that conversion for later.

Pay no attention to the WAR file just yet. Until the project is done,
the service is better off simply being deployed to glassfish directly
through adding the project to a server instance within eclipse. This
takes advantage of the debugging attachment ability as well as the 'Hot
Code Replacement' for seeing changes more instantaneously.