Service Layer - GET Appointments & POST Appointments
The service accurately responds with XML of all the appointments to spec
and responds with correct XML for a successful appointment creation.

XStream converters have proved very useful in this process. However, the
layers created at server runtime are not beans! Therefore, nothing
should be expected to persist!

Probably something to run by dean.