CORS for all Service Endpoints & IVY
Fully switched to Ivy dependency management.
No more excessive lib folder. (This was mainly in an effort to sort out
a bug where the connections to derby kept failing. I thought that maybe
it was just because we were using an older version)

All service endpoints return a response now so the CORS Header can be
present for any AJAX calls.